refactoring quest chain page for different versions
